+++ src/main/java/roomescape/common/exception/InternalServerErrorException.java
@@ -1,0 +1,10 @@
+package roomescape.common.exception;
+
+import roomescape.common.exception.vo.ErrorCode;
+
+public class InternalServerErrorException extends CustomException {
+
+    public InternalServerErrorException() {
+        super(ErrorCode.SERVER_ERROR);
+    }
+}
