+++ src/main/java/roomescape/infrastructure/dto/response/ConfirmPaymentResponse.java
@@ -1,0 +1,10 @@
+package roomescape.infrastructure.dto.response;
+
+import roomescape.infrastructure.dto.PaymentFailure;
+
+public record ConfirmPaymentResponse(
+        Integer totalAmount,
+        String paymentKey,
+        PaymentFailure failure
+) {
+}
