+++ src/test/java/roomescape/reservation/service/CreateReservationServiceTest.java
@@ -16,7 +16,7 @@ import roomescape.common.exception.NotFoundException;
 import roomescape.config.RestClientConfig;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
-import roomescape.payment.service.TossPaymentService;
+import roomescape.infrastructure.TossPaymentClient;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.service.dto.request.ReservationCreateRequest;
 import roomescape.reservation.service.dto.response.ReservationResponse;
@@ -33,7 +33,7 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 @ActiveProfiles("test")
 @DataJpaTest
-@Import({CreateReservationService.class, RestClientConfig.class, TossPaymentService.class})
+@Import({CreateReservationService.class, RestClientConfig.class, TossPaymentClient.class})
 class CreateReservationServiceTest {
 
     private final LocalDateTime now = LocalDateTime.now();
